Add action for removal of sections with additional info/metadata
We split element removal into yet another type of action. The
reasoning behind this is that different clients have different
needs:

When taking screenshots, we might want to remove e.g. location
info, because it's frequently badly formatted visual noise.

On the other hand, if we are crawling for textual content, we'll
want to remove cookie popus and other info boxes, but location info
remains relevant.